requirements.txt
Flask
psycopg2-binary
python-dotenv
flask_sqlalchemy
flask_migrate
pyjwt